Added userConfig Class
